requirements.txt
python-chess
tensorflow-gpu>=1.7.0

